Create autocmd User Vebugger_DebuggerActive
This is spcecially useful if you want to activate some keymapping only
during the debug session. This could be achieved with something like
this on your `.vimrc`:

let s:debug_mappings = [
  \ {'key': 'n', 'func': ':VBGstepOver'},
  \ {'key': 's', 'func': ':VBGstepIn'},
  \ {'key': 'c', 'func': ':VBGcontinue'},
  \ {'key': 'S', 'func': ':VBGstepOut'},
  \ {'key': 'b', 'func': ':VBGtoggleBreakpointThisLine'},
  \ {'key': 'B', 'func': ':VBGclearBreakpoints'},
  \ {'key': 'k', 'func': ':call KillJavaDebugger()'},
  \ ]

function! InjectDebugMappings()
  for map in s:debug_mappings
    exe 'nnoremap <silent> <buffer> ' . map['key'] . ' ' . map['func'] . '<CR>'
  endfor
endfunction

augroup Vebugger
  autocmd!
  autocmd User Vebugger_DebuggerActive call InjectDebugMappings()
augroup END

Is this case, the function `KillJavaDebugger()` is responsible to remove
the mappings. But this could be achieved with another User autocmd named
`Vebugger_DebuggerInactive`, for example.